Note migration
